re-structured files to remove main.py
