+++ new_src/data.py
@@ -1,11 +1,14 @@
 import os
-from typing import Union, Callable
+from typing import Union, Callable, Optional
 
+import matplotlib.pyplot as plt
+from matplotlib.axes import Axes
 import torch
-import torch.nn as nn
-from torchvision import datasets
-from torchvision import transforms
+from torch import nn, Tensor
+from torchvision import datasets, transforms
 from torch.utils.data import DataLoader, TensorDataset, Dataset
+
+from new_src.architectures import AutoEncoder, VarAutoEncoder
 
 
 class TransposeTransform(nn.Module):
@@ -56,91 +59,82 @@ def load_EMNIST(
     return dataloader
 
 
+def encode_dataset(
+    data: Union[Dataset, DataLoader],
+    autoencoder: Union[AutoEncoder, VarAutoEncoder],
+    save_path: Optional[str] = None,
+    root: str = "data",
+    batch_size: int = 128,
+) -> tuple[Tensor, Tensor]:
+    """
+    Creates dataset with features (autoencoder.encode(x), y)
+    and saves it under root/save_path/ as z.pt, y.pt
+    """
+    if not isinstance(data, DataLoader):
+        dataloader = DataLoader(data, batch_size=batch_size, shuffle=False)
+    else:
+        dataloader = data
+    device = next(autoencoder.parameters()).device
+
+    n_batch = len(dataloader.dataset)  # type: ignore
+    data_z = torch.empty(n_batch, *autoencoder.z_shape)
+    data_y = torch.empty(n_batch)
+    with torch.no_grad():
+        idx = 0
+        for x, y in dataloader:
+            batches = x.shape[0]
+            z = autoencoder.encode(x.to(device)).detach().cpu()
+            data_z[idx: idx + batches] = z
+            data_y[idx: idx + batches] = y
+            idx += batches
+
+    if save_path:
+        path_z = os.path.join(root, save_path, "z.pt")
+        path_y = os.path.join(root, save_path, "y.pt")
+        torch.save(data_z, path_z)
+        torch.save(data_y, path_y)
+
+    return data_z, data_y
+
+
 def load_TensorDataset(
-    root: str,
+    save_path: str,
     batch_size: int = 128,
     shuffle: bool = False,
     num_workers: int = 0,
+    root: str = "data",
+    names: list[str] = ["z.pt", "y.pt"]
 ) -> DataLoader:
     """
-    Loads file stored in .pt files.
-    features data is in root/x.pt
-    labels data is in root/y.pt
+    Loads specified files in root/save_path/
 
     Args:
-        root: path (from directory) to folder with dataset
+        save_path: path from data folder to files folder
         batch_size: batch_size for dataloader
         shuffle: whether to shuffle the dataloader
         num_workers: parallelize computation. If num_workers > 0, then
             pin_memory=True and persistent_workers=True
+        root: path from directory to data folder
+        names: list of file names to retrieve
 
     Returns:
         dataloader with the saved dataset
     """
     torch.cuda.empty_cache()
-    # define arguments
     workers_args = {
         "num_workers": num_workers,
         "pin_memory": True,
         "persistent_workers": True,
     } if num_workers > 0 else {}
-    # load dataset
-    path_x = os.path.join(root, "x.pt")
-    path_y = os.path.join(root, "y.pt")
-    tensor_x = torch.load(path_x)
-    tensor_y = torch.load(path_y)
-    dataset = TensorDataset(tensor_x, tensor_y)
-    # return dataloader
+
+    data = []
+    for name in names:
+        path = os.path.join(root, save_path, name)
+        tensor = torch.load(path)
+        data.append(tensor)
+    dataset = TensorDataset(*data)
+
     dataloader = DataLoader(
         dataset, batch_size=batch_size, shuffle=shuffle, **workers_args,
     )
     return dataloader
-
-
-def process_dataset(
-    data: Union[Dataset, DataLoader],
-    save_path: str,
-    fn_x: Callable,
-    fn_y: Callable,
-    shape_x: tuple[int],
-    shape_y: tuple[int],
-    batch_size: int = 128,
-    device: str = 'cuda',
-) -> None:
-    """
-    Process a dataset according to given functions.
-    Store tensors in save_path/x.pt and save_path/y.pt
-
-    Args:
-        data: original dataset/dataloader to process
-        save_path: relative path to save tensors
-        fn_x: function to apply to batched inputs (on device)
-        fn_y: function to apply to batched labels
-        shape_x: shape of a processed x tensor
-        shape_y: shape of a processed y tensor
-        batch_size: batch_size to use if data is a Dataset
-    """
-    # make dataloader
-    if not isinstance(data, DataLoader):
-        dataloader = DataLoader(data, batch_size=batch_size, shuffle=False)
-    else:
-        dataloader = data
-    # evaluation loop
-    newdata_x = torch.zeros(len(dataloader.dataset), *shape_x)  # type: ignore
-    newdata_y = torch.zeros(len(dataloader.dataset), *shape_y)  # type: ignore
-    with torch.no_grad():
-        idx = 0
-        for x, y in dataloader:
-            B = x.shape[0]
-            x, y = x.to(device), y.to(device)
-            new_x = fn_x(x, y).detach().cpu()
-            new_y = fn_y(x, y).detach().cpu()
-            newdata_x[idx: idx + B] = new_x
-            newdata_y[idx: idx + B] = new_y
-            idx += B
-    # save tensors
-    if save_path:
-        path_x = os.path.join(save_path, "x.pt")
-        path_y = os.path.join(save_path, "y.pt")
-        torch.save(newdata_x, path_x)
-        torch.save(newdata_y, path_y)
